feat: geofence 알림 구현 (#23)
* geofence초안 추가

* geofence 할 일 알림 추가 (enter,exit)

* geofence클래스 타입오류 해결 / 맵 위치 수정 후 geofence 변경 추가

* geofence오류 수정/ place isEnter추가

* 오류수정1

* 오류수정2

* 06-18 05:28

* 리시버- 요일, iscompleted 관련 추가

* 리시버- 요일, iscompleted 관련 수정

* 알림 관련 수정

* refactor: Geofence 수정사항 구현 (#30)

* refactor: Todo, Place id 자료형 Long으로 변경

* refactor: XML Empty body

* feat: Geofence 수정, 장소 추가 시 Geofence 등록 및 장소 삭제 시 Geofence 제거하도록 구현

---------

+++ app/src/main/java/org/konkuk/placelist/main/AddPlaceListener.kt
@@ -1,6 +1,6 @@
 package org.konkuk.placelist.main
 
 interface AddPlaceListener {
-    fun addPlace(id: Int, name: String, latitude: String, longitude: String, radius: Float)
+    fun addPlace(id: Long, name: String, latitude: String, longitude: String, radius: Float)
 
 }
